Working code with comments

+++ Gym/include/Movati.hpp
@@ -1,12 +1,12 @@
 #ifndef __INHERITANCE__MOVATI__HPP__
 #define __INHERITANCE__MOVATI__HPP__
 
-#include "Gym.hpp"
+#include <Gym.hpp>
 
 class Movati : public Gym {
 public:
-    float cardio() const override;
-    float lift() const override;
+    double cardio() const override;
+    double lift() const override;
 };
 
 #endif
