+++ Gym/src/Crunch.cpp
@@ -1,10 +1,10 @@
 
-#include "Crunch.hpp"
+#include <Crunch.hpp> //includes the header file to allow access to its functions
 
-float Crunch::cardio() const {
+double Crunch::cardio() const { // instance cardio from crunch that returns a float of increase percentage 
     return -3.0f;
 }
 
-float Crunch::lift() const {
+double Crunch::lift() const { // instance lfit from crunch that returns a float of increase percentage 
     return 1.0f;
 }
